Add documentation how to migrate to Amazon EKS add-ons
Since those components are highly related to the K8s version and managing
them from a central point makes sense long term. AWS is also aware of
which components are deployed, should this ever be relevant for the
EKS upgrade process.

Only the versions have to be managed, so less overhead to take care of
updating the Project Syn component eks-addon-manager.